disabled operations when in deleteAll mode

+++ src/components/Item.js
@@ -3,20 +3,28 @@ import {useState} from "react";
 function Item(props) {
     const [prevText, setPrevText] = useState(props.text)
     return (
-        <div className={!props.isCompleted ?
-            props.editID === props.id ? `selectedListItem` : `listItem`
-            :
-            `completeItem`}
-             onClick={() => props.onClick(props.id)}>
+        <div className={
+            !props.isCompleted ?
+                props.editID === props.id ? `selectedListItem` : `listItem`
+                :
+                `completeItem`}
+                onClick={() => {
+                    if (props.deleteConfirm) return;
+                    props.onClick(props.id)
+                }}>
+
+            {/*checkbox*/}
             <input type="checkbox" name="buyBook"
                    checked={props.isCompleted}
                    onClick={(e) => e.stopPropagation()}
                    onChange={() => {
                        if (props.editID !== false) return
+                       if (props.deleteConfirm) return;
                        props.onChange(props.id, "isCompleted", !props.isCompleted);
                    }
                    }
             />
+
             {props.editID === props.id ?
                 <input type="text"
                        className="rename"
@@ -27,11 +35,14 @@ function Item(props) {
                 :
                 <div className="item"> {props.text} </div>
             }
+
+            {/* Delete button */}
             <div className="buttonGroup" onClick={(e) => e.stopPropagation()}>
                 <button
                     className="deleteButton"
                     onClick={() => {
                         if (props.editID !== false) return
+                        if (props.deleteConfirm) return;
                         props.onDelete(props.id)
                     }}
                 > X
